ticket-app/requirements.txt
blinker
certifi
charset-normalizer
click
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-table
Flask
idna
importlib_metadata
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
nest-asyncio
numpy
packaging
pandas
plotly
PyMySQL
python-dateutil
python-dotenv
pytz
requests
retrying
six
tenacity
typing_extensions
tzdata
urllib3
Werkzeug
zipp
